adjust ValidSubmissionTimestampValidator and use Integer

+++ common/persistence/src/main/java/app/coronawarn/server/common/persistence/domain/validation/ValidSubmissionTimestampValidator.java
@@ -1,5 +1,3 @@
-
-
 package app.coronawarn.server.common.persistence.domain.validation;
 
 import java.time.Instant;
@@ -8,13 +6,13 @@ import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 
 public class ValidSubmissionTimestampValidator
-    implements ConstraintValidator<ValidSubmissionTimestamp, Long> {
+    implements ConstraintValidator<ValidSubmissionTimestamp, Integer> {
 
   public static final long SECONDS_PER_HOUR = TimeUnit.HOURS.toSeconds(1);
 
   @Override
-  public boolean isValid(Long submissionTimestamp, ConstraintValidatorContext constraintValidatorContext) {
+  public boolean isValid(Integer submissionTimestamp, ConstraintValidatorContext constraintValidatorContext) {
     long currentHoursSinceEpoch = Instant.now().getEpochSecond() / SECONDS_PER_HOUR;
-    return submissionTimestamp >= 0L && submissionTimestamp <= currentHoursSinceEpoch;
+    return submissionTimestamp >= 0 && submissionTimestamp <= currentHoursSinceEpoch;
   }
 }
